main에서 mypage 이동 연결
